requirements.txt
streamlit==1.40.2
streamlit-webrtc==0.47.0
opencv-python
pillow==11.0.0
numpy==1.26.4
requests==2.32.3
ultralytics>=8.0.0
torch==2.2.2
torchvision==0.17.2